Alpha 0.1 Pre-release 8
Improved code and some actions. Fixes #1. Fixed other various bugs.

+++ src/sune/ssp/util/Resource.java
@@ -10,10 +10,18 @@ import javafx.scene.image.Image;
 public class Resource {
 	
 	public static URL resource(String path) {
+		if(path == null || path.isEmpty()) {
+			throw new IllegalArgumentException(
+				"Path cannot be null or empty!");
+		}
 		return Resource.class.getResource(path);
 	}
 	
 	public static InputStream stream(String path) {
+		if(path == null || path.isEmpty()) {
+			throw new IllegalArgumentException(
+				"Path cannot be null or empty!");
+		}
 		return Resource.class.getResourceAsStream(path);
 	}
 	
@@ -22,14 +30,17 @@ public class Resource {
 	}
 	
 	public static File file(String path) {
-		return file("temp_", path);
+		return file(path, "temp_", ".tmp");
 	}
 	
-	public static File file(String prefix, String path) {
+	public static File file(String path, String prefix, String suffix) {
+		if(path == null || path.isEmpty()) {
+			throw new IllegalArgumentException(
+				"Path cannot be null or empty!");
+		}
 		try {
-			File file = File.createTempFile(prefix, ".tmp");
+			File file = File.createTempFile(prefix, suffix);
 			file.deleteOnExit();
-			
 			try(InputStream is = stream(path)) {
 				try(FileOutputStream os = new FileOutputStream(file)) {
 					byte[] buffer = new byte[8192];
@@ -37,13 +48,11 @@ public class Resource {
 					int read;
 					while((read = is.read(buffer)) != -1)
 						os.write(buffer, 0, read);
-					
 					return file;
 				}
 			}
 		} catch(Exception ex) {
 		}
-		
 		return null;
 	}
